Refactors functionality distribution across modules
